update them sua xoa tìm kiem các bang

+++ btbongda/src/main/java/com/example/btbongda/repository/MatchRepository.java
@@ -3,5 +3,5 @@ package com.example.btbongda.repository;
 import com.example.btbongda.entity.MatchInfors;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchRepository extends JpaRepository<MatchInfors, Integer> {
+public interface MatchRepository extends JpaRepository<MatchInfors, Long> {
 }
